q3map2/main: fix: warning: implicit conversion from int to vec_t (aka float) changes value from 2147483647 to 2147483648

+++ tools/quake3/q3map2/main.c
@@ -43,7 +43,7 @@
  */
 
 vec_t Random( void ){
-	return (vec_t) rand() / RAND_MAX;
+	return (vec_t) rand() / (float) RAND_MAX;
 }
